DROOLS-4582: Fixed cell reselection
